backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
alembic
asyncpg
aiosqlite
redis
httpx
python-dotenv
google-cloud-storage
google-cloud-translate
google-cloud-vision
google-cloud-texttospeech
google-cloud-speech
google-auth
google-generativeai
librosa
